backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
qdrant-client
groq
python-dotenv
requests
supabase==2.3.0
python-multipart
